+++ core-ng/src/main/java/core/framework/internal/db/DatabaseImpl.java
@@ -112,9 +112,6 @@ public final class DatabaseImpl implements Database {
             // mysql with ssl has overhead, usually we ensure security on arch level, e.g. gcloud sql proxy or firewall rule
             if (index == -1 || url.indexOf("useSSL=", index + 1) == -1) properties.setProperty(PropertyKey.useSSL.getKeyName(), "false");
             // refer to https://dev.mysql.com/doc/connector-j/8.0/en/connector-j-reference-charsets.html
-            // without characterEncoding, mysql driver auto detects charset based on server side var character_set_server
-            // and only character_set_server=utf8mb4 will map characterEncoding to utf-8
-            // here use utf-8 as default value, and support to override by jdbc url param
             if (index == -1 || url.indexOf("characterEncoding=", index + 1) == -1) properties.setProperty(PropertyKey.characterEncoding.getKeyName(), "utf-8");
         }
         return properties;